feat: multiplayer communication
Implemented a message registry (MessageLog) to hold on incoming messages. Messages
are stored in a key-value structure where key is computed from the
message attributes `$turn-$event-$sender-$recipient`, where `$recipient`
can be optional, and the stored value is the entire message. Using this
utility class we can poll for the correct messages for each step of the
gameplay without worrying too much of race conditions. Messages will
arrive sooner or later and they will be polled when neeeded.